--- src/app/confirm/page.module.css
@@ -1,144 +1,0 @@
-/* src/app/confirm/page.module.css */
-.wrap {
-  display: flex;
-  flex-direction: column;
-  min-height: 100dvh;
-  --mobile-fixed-cta-clearance: calc(96px + env(safe-area-inset-bottom));
-}
-
-.header {
-  background: var(--navy);
-  padding: 20px 20px 16px;
-}
-.headerTop {
-  display: flex;
-  align-items: baseline;
-  justify-content: space-between;
-  margin-bottom: 6px;
-}
-.title { font-size: 19px; font-weight: 800; letter-spacing: -0.5px; color: #fff; }
-.count { font-size: 13px; font-weight: 600; color: rgba(255,255,255,.55); }
-.hint { font-size: 13px; color: rgba(255,255,255,.75); line-height: 1.5; }
-
-.cashSection {
-  padding: 16px 16px 0;
-}
-
-.cashHint {
-  display: block;
-  margin-top: 6px;
-  font-size: 13px;
-  line-height: 1.5;
-  color: var(--text-2);
-}
-
-.cashInputWrap {
-  display: block;
-  padding: 16px;
-  border: 1px solid var(--border);
-  border-radius: 16px;
-  background: #fff;
-}
-
-.cashLabel {
-  display: block;
-  font-size: 14px;
-  font-weight: 700;
-  color: var(--text-1);
-}
-
-.cashInputRow {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  margin-top: 10px;
-}
-
-.cashInput {
-  width: 100%;
-  min-width: 0;
-  padding: 14px 16px;
-  border: 1px solid var(--border);
-  border-radius: 12px;
-  font-size: 16px;
-  font-weight: 700;
-  color: var(--text-1);
-  background: #fff;
-}
-
-.cashInput:focus {
-  outline: none;
-  border-color: var(--navy);
-  box-shadow: 0 0 0 3px rgba(15, 23, 42, 0.08);
-}
-
-.cashUnit {
-  flex-shrink: 0;
-  font-size: 14px;
-  font-weight: 700;
-  color: var(--text-2);
-}
-
-.scroll {
-  flex: 1;
-  overflow-y: auto;
-  padding: 12px 16px var(--mobile-fixed-cta-clearance);
-  scroll-padding-bottom: var(--mobile-fixed-cta-clearance);
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.dupNotice {
-  background: var(--amber-bg);
-  border: 1px solid var(--amber-border);
-  border-radius: 12px;
-  padding: 14px 16px;
-  font-size: 13px;
-  color: #92400E;
-  line-height: 1.55;
-}
-.dupNotice strong { font-weight: 700; }
-
-.empty {
-  text-align: center;
-  padding: 48px 24px;
-  color: var(--text-2);
-  font-size: 14px;
-  line-height: 1.6;
-}
-
-/* 768px+ 테이블 뷰 */
-.tableWrap { display: none; }
-
-.table {
-  width: 100%;
-  border-collapse: collapse;
-  font-size: 14px;
-}
-.thead tr { border-bottom: 2px solid var(--border); }
-.th {
-  padding: 10px 12px 10px 0;
-  font-size: 11px;
-  font-weight: 700;
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-  color: var(--text-3);
-  text-align: left;
-  white-space: nowrap;
-}
-.thNum { text-align: right; }
-
-@media (min-width: 768px) {
-  .scroll { display: none; }
-  .tableWrap {
-    display: block;
-    flex: 1;
-    overflow-y: auto;
-    padding: 16px 32px 24px;
-  }
-
-  .cashSection {
-    padding: 20px 32px 0;
-  }
-}
