--- src/app/diagnosis/page.module.css
@@ -1,393 +1,0 @@
-/* src/app/diagnosis/page.module.css */
-.wrap {
-  display: flex;
-  flex-direction: column;
-  min-height: 100dvh;
-  background: var(--bg);
-}
-
-.header {
-  padding: 0;
-}
-
-.headerCard {
-  overflow: hidden;
-  margin: 16px 16px 0;
-  border-radius: 20px;
-  background: var(--navy);
-}
-
-.headerTopRow {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 12px;
-  padding: 16px 18px 0;
-}
-
-.date {
-  font-size: 11px;
-  font-weight: 500;
-  color: rgba(255, 255, 255, 0.4);
-}
-
-.retryBtn {
-  border: none;
-  border-radius: 20px;
-  background: rgba(255, 255, 255, 0.1);
-  padding: 5px 10px;
-  font-size: 11px;
-  font-weight: 700;
-  color: rgba(255, 255, 255, 0.65);
-}
-
-.retryBtn:active {
-  background: rgba(255, 255, 255, 0.18);
-}
-
-.headerBody {
-  padding: 16px 18px 0;
-}
-
-.headline {
-  margin-bottom: 6px;
-  font-size: 12px;
-  font-weight: 500;
-  color: rgba(255, 255, 255, 0.45);
-}
-
-.bigNum {
-  margin-bottom: 4px;
-  font-size: 52px;
-  font-weight: 900;
-  letter-spacing: -2px;
-  line-height: 1;
-  color: #fff;
-}
-
-.bigNum em {
-  margin-left: 6px;
-  font-style: normal;
-  font-size: 20px;
-  font-weight: 700;
-  color: rgba(255, 255, 255, 0.85);
-  vertical-align: 8px;
-}
-
-.healthyTitle {
-  margin-bottom: 6px;
-  font-size: 30px;
-  font-weight: 900;
-  letter-spacing: -0.05em;
-  line-height: 1.04;
-  color: #fff;
-}
-
-.subline {
-  margin-bottom: 14px;
-  font-size: 13px;
-  font-weight: 600;
-  color: rgba(255, 255, 255, 0.72);
-}
-
-.healthScore {
-  display: inline-flex;
-  align-items: center;
-  gap: 6px;
-  border-radius: 999px;
-  background: rgba(255, 255, 255, 0.1);
-  padding: 5px 12px;
-  margin-bottom: 20px;
-}
-
-.healthScoreDot {
-  width: 7px;
-  height: 7px;
-  border-radius: 50%;
-  background: #f2ae34;
-}
-
-.healthScoreLabel {
-  font-size: 12px;
-  font-weight: 500;
-  color: rgba(255, 255, 255, 0.8);
-}
-
-.healthScoreValue {
-  font-size: 16px;
-  font-weight: 800;
-  letter-spacing: -0.03em;
-  color: #fff;
-}
-
-.allocationPanel {
-  margin: 0 12px 12px;
-  border-radius: 14px;
-  background: rgba(0, 0, 0, 0.16);
-  padding: 14px 16px;
-}
-
-.allocationPanelHeader {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 12px;
-  margin-bottom: 10px;
-}
-
-.allocationPanelTitle {
-  font-size: 11px;
-  font-weight: 600;
-  letter-spacing: 0.5px;
-  color: rgba(255, 255, 255, 0.35);
-}
-
-.allocationPanelMeta {
-  font-size: 10px;
-  color: rgba(255, 255, 255, 0.25);
-}
-
-.targetError {
-  margin-top: 10px;
-  padding: 0 4px;
-  font-size: 12px;
-  font-weight: 700;
-  color: #ffb4b4;
-}
-
-.body {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  padding: 10px 16px 112px;
-}
-
-.improvementSection {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.sectionLabel {
-  padding: 10px 4px 2px;
-  font-size: 11px;
-  font-weight: 800;
-  letter-spacing: 0.1em;
-  text-transform: uppercase;
-  color: var(--text-3);
-}
-
-.problemList {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.improvementBtn,
-.signalBtn,
-.explainBtn {
-  width: 100%;
-  border-radius: var(--radius-card);
-  background: var(--surface);
-  text-align: left;
-}
-
-.improvementBtn,
-.signalBtn {
-  border: 1px solid var(--border);
-  padding: 16px 18px;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 14px;
-}
-
-.improvementBtn:active,
-.signalBtn:active,
-.explainBtn:active {
-  background: #f7f8fc;
-}
-
-.improvementBtnText,
-.signalBtnText {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-  min-width: 0;
-}
-
-.improvementBtnTitle,
-.signalBtnTitle {
-  font-size: 14px;
-  font-weight: 800;
-  letter-spacing: -0.02em;
-  color: var(--text-1);
-}
-
-.improvementBtnMeta,
-.signalBtnMeta {
-  font-size: 12px;
-  font-weight: 600;
-  line-height: 1.5;
-  color: var(--text-2);
-}
-
-.improvementBtnArrow,
-.signalBtnArrow {
-  font-size: 16px;
-  font-weight: 800;
-  color: var(--navy);
-  flex-shrink: 0;
-}
-
-.explainBtn {
-  border: 1px solid var(--border);
-  padding: 16px 18px;
-  font-size: 14px;
-  font-weight: 700;
-  color: var(--text-2);
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  letter-spacing: -0.02em;
-}
-
-.explainIcon {
-  font-size: 11px;
-  transition: transform 0.2s;
-}
-
-.explainOpen .explainIcon {
-  transform: rotate(180deg);
-}
-
-.explainBody {
-  margin-top: -10px;
-  border: 1px solid var(--border);
-  border-top: none;
-  border-radius: 0 0 var(--radius-card) var(--radius-card);
-  background: var(--surface);
-  padding: 18px;
-  font-size: 13px;
-  line-height: 1.7;
-  color: var(--text-2);
-}
-
-.explainBody p + p {
-  margin-top: 12px;
-}
-
-.explainBody button {
-  border: none;
-  background: none;
-  padding: 0;
-  font-size: 13px;
-  color: var(--navy);
-  text-decoration: underline;
-}
-
-.mbtiCard {
-  border-radius: var(--radius-card);
-  padding: 20px 18px;
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-  box-shadow: 0 6px 24px rgba(28, 43, 94, 0.25);
-}
-
-.mbtiEyebrow {
-  font-size: 10px;
-  font-weight: 700;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-  color: rgba(255, 255, 255, 0.45);
-}
-
-.mbtiTop {
-  display: flex;
-  align-items: center;
-  gap: 14px;
-}
-
-.mbtiEmojiBox {
-  width: 50px;
-  height: 50px;
-  flex-shrink: 0;
-  border-radius: 14px;
-  background: rgba(255, 255, 255, 0.12);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 24px;
-}
-
-.mbtiCode {
-  font-size: 30px;
-  font-weight: 900;
-  letter-spacing: -0.05em;
-  line-height: 1;
-  color: #fff;
-}
-
-.mbtiName {
-  margin-top: 3px;
-  font-size: 12px;
-  font-weight: 600;
-  color: rgba(255, 255, 255, 0.52);
-}
-
-.mbtiTagline {
-  font-size: 14px;
-  font-weight: 700;
-  line-height: 1.5;
-  color: #fff;
-}
-
-.mbtiDesc {
-  font-size: 12.5px;
-  line-height: 1.65;
-  color: rgba(255, 255, 255, 0.72);
-}
-
-.mbtiGroup_analyst {
-  background: linear-gradient(140deg, #1c2b5e 0%, #2d479b 100%);
-}
-
-.mbtiGroup_diplomat {
-  background: linear-gradient(140deg, #4c1d95 0%, #7c3aed 100%);
-}
-
-.mbtiGroup_sentinel {
-  background: linear-gradient(140deg, #064e3b 0%, #059669 100%);
-}
-
-.mbtiGroup_explorer {
-  background: linear-gradient(140deg, #7c2d12 0%, #ea580c 100%);
-}
-
-.shareBtn {
-  align-self: stretch;
-  border: 1.5px solid rgba(255, 255, 255, 0.2);
-  border-radius: 12px;
-  background: rgba(255, 255, 255, 0.12);
-  padding: 13px 16px;
-  font-size: 13px;
-  font-weight: 700;
-  color: #fff;
-}
-
-.shareBtn:active {
-  background: rgba(255, 255, 255, 0.22);
-}
-
-.disclaimer {
-  padding: 4px 8px;
-  font-size: 11px;
-  line-height: 1.6;
-  text-align: center;
-  color: var(--text-3);
-}
-
-.disclaimer p + p {
-  margin-top: 4px;
-}
