--- src/app/market/page.module.css
@@ -1,336 +1,0 @@
-.wrap {
-  min-height: 100dvh;
-  background: #edeef3;
-}
-
-.page {
-  min-height: 100dvh;
-  padding-bottom: 88px;
-}
-
-.header {
-  padding: 20px 18px 14px;
-}
-
-.eyebrow {
-  margin-bottom: 3px;
-  font-size: 12px;
-  font-weight: 500;
-  color: #999;
-}
-
-.title {
-  font-size: 20px;
-  font-weight: 800;
-  letter-spacing: -0.04em;
-  color: #111;
-}
-
-.description {
-  margin-top: 5px;
-  font-size: 12.5px;
-  line-height: 1.55;
-  color: #aaa;
-}
-
-.content {
-  padding: 0 0 16px;
-}
-
-.heroCard,
-.stateCard,
-.skeletonCard {
-  margin: 0 16px 12px;
-  background: #fff;
-  border-radius: 18px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.06);
-}
-
-.heroCard {
-  padding: 16px 18px 14px;
-  display: grid;
-  gap: 10px;
-}
-
-/* KPI: 진입 매력도 */
-.entryBlock {
-  display: grid;
-  gap: 4px;
-}
-
-.kpiEyebrow {
-  font-size: 10px;
-  font-weight: 700;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-  color: #aaa;
-}
-
-.kpiRow {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.kpiScore {
-  display: flex;
-  align-items: baseline;
-  gap: 2px;
-}
-
-.kpiScoreNum {
-  font-size: 38px;
-  font-weight: 900;
-  letter-spacing: -0.05em;
-  line-height: 1;
-  color: #1c2b5e;
-  font-variant-numeric: tabular-nums;
-}
-
-.kpiScoreUnit {
-  font-size: 14px;
-  font-weight: 700;
-  color: #ccc;
-}
-
-.kpiLabel {
-  font-size: 15px;
-  font-weight: 800;
-  letter-spacing: -0.03em;
-  color: #1c2b5e;
-}
-
-.kpiSummary {
-  font-size: 12px;
-  line-height: 1.6;
-  color: #888;
-}
-
-/* Sub KPI: 시장 건강도 */
-.healthChip {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 9px 12px;
-  border-radius: 10px;
-  background: #f4f6ff;
-  border: 1px solid #e0e4f0;
-}
-
-.healthEyebrow {
-  font-size: 10px;
-  font-weight: 700;
-  letter-spacing: 0.06em;
-  text-transform: uppercase;
-  color: #aaa;
-  flex-shrink: 0;
-}
-
-.healthLabel {
-  font-size: 13px;
-  font-weight: 800;
-  color: #333;
-  flex: 1;
-}
-
-.healthScore {
-  font-size: 12px;
-  font-weight: 700;
-  color: #aaa;
-  font-variant-numeric: tabular-nums;
-}
-
-/* 투자 시사점 */
-.implication {
-  display: grid;
-  gap: 4px;
-  padding: 10px 12px;
-  border-radius: 10px;
-  background: rgba(28, 43, 94, 0.04);
-}
-
-.implicationLabel {
-  font-size: 10px;
-  font-weight: 800;
-  letter-spacing: 0.06em;
-  text-transform: uppercase;
-  color: rgba(28, 43, 94, 0.5);
-}
-
-.implicationText {
-  font-size: 12px;
-  line-height: 1.6;
-  color: #555;
-}
-
-.sectionHeader {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 12px;
-  padding: 0 18px 10px;
-}
-
-.sectionEyebrow {
-  font-size: 12px;
-  font-weight: 700;
-  color: #888;
-}
-
-.refreshButton,
-.inlineRetry {
-  border: none;
-  background: rgba(28, 43, 94, 0.08);
-  color: #1c2b5e;
-  min-height: 34px;
-  padding: 0 12px;
-  border-radius: 999px;
-  font-size: 11px;
-  font-weight: 700;
-}
-
-.cardList {
-  display: grid;
-  gap: 0;
-}
-
-.metaRow {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  padding: 6px 16px 0;
-}
-
-.metaBadge {
-  border-radius: 999px;
-  padding: 6px 10px;
-  background: rgba(255, 255, 255, 0.8);
-  border: 1px solid #e3e5ea;
-  font-size: 11px;
-  font-weight: 700;
-  color: #7b7f89;
-}
-
-.source {
-  margin: 6px 16px 0;
-  font-size: 11px;
-  line-height: 1.7;
-  text-align: center;
-  color: #bbb;
-}
-
-.stateCard {
-  padding: 18px;
-  font-size: 14px;
-  line-height: 1.6;
-  color: var(--text-2);
-}
-
-.inlineRetry {
-  margin-top: 12px;
-}
-
-.skeletonRow {
-  display: flex;
-  align-items: center;
-}
-
-.skeletonKpiEyebrow,
-.skeletonKpiRow,
-.skeletonKpiScore,
-.skeletonKpiLabel,
-.skeletonHealthChip,
-.skeletonLine,
-.skeletonTitleShort,
-.skeletonIcon,
-.skeletonBadge {
-  background: linear-gradient(90deg, #f2f3f7, #e7e9ef, #f2f3f7);
-  background-size: 200% 100%;
-  animation: shimmer 1.4s linear infinite;
-}
-
-.skeletonKpiEyebrow {
-  width: 64px;
-  height: 10px;
-  border-radius: 999px;
-}
-
-.skeletonKpiRow {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  background: none;
-  animation: none;
-}
-
-.skeletonKpiScore {
-  width: 72px;
-  height: 38px;
-  border-radius: 8px;
-}
-
-.skeletonKpiLabel {
-  width: 90px;
-  height: 16px;
-  border-radius: 999px;
-}
-
-.skeletonLine {
-  width: 180px;
-  height: 11px;
-  border-radius: 999px;
-}
-
-.skeletonHealthChip {
-  width: 100%;
-  height: 36px;
-  border-radius: 10px;
-}
-
-.skeletonCard {
-  padding: 14px 16px;
-  margin-bottom: 8px;
-}
-
-.skeletonRow {
-  gap: 12px;
-}
-
-.skeletonIcon {
-  width: 36px;
-  height: 36px;
-  border-radius: 10px;
-  flex-shrink: 0;
-}
-
-.skeletonText {
-  flex: 1;
-}
-
-.skeletonTitleShort {
-  width: 74px;
-  height: 13px;
-  margin-bottom: 8px;
-}
-
-.skeletonLine {
-  width: 180px;
-  height: 11px;
-}
-
-.skeletonBadge {
-  width: 44px;
-  height: 22px;
-  border-radius: 999px;
-  flex-shrink: 0;
-}
-
-@keyframes shimmer {
-  from {
-    background-position: 200% 0;
-  }
-
-  to {
-    background-position: -200% 0;
-  }
-}
