--- src/app/page.module.css
@@ -1,126 +1,0 @@
-/* src/app/page.module.css */
-.wrap { display: flex; flex-direction: column; min-height: 100dvh; }
-
-.body {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 48px 24px 40px;
-  gap: 0;
-}
-
-.intro { text-align: center; margin-bottom: 40px; }
-.headline {
-  font-size: 28px;
-  font-weight: 800;
-  letter-spacing: -0.8px;
-  line-height: 1.28;
-  margin-bottom: 12px;
-}
-.sub { font-size: 14px; color: var(--text-2); line-height: 1.6; }
-
-.dropZone {
-  width: 100%;
-  background: var(--surface);
-  border: 1.5px solid var(--border-2);
-  border-radius: 18px;
-  padding: 36px 24px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 14px;
-  cursor: pointer;
-  text-align: center;
-  margin-bottom: 20px;
-  transition: border-color 0.15s, box-shadow 0.15s;
-}
-.dropZone:active {
-  border-color: var(--navy);
-  box-shadow: 0 0 0 3px rgba(26,35,126,.08);
-}
-
-.iconWrap {
-  width: 52px;
-  height: 52px;
-  background: var(--navy);
-  border-radius: 14px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-.zoneTitle { font-size: 16px; font-weight: 700; letter-spacing: -0.3px; }
-.zoneDesc { font-size: 13px; color: var(--text-2); line-height: 1.55; }
-
-.privacy {
-  font-size: 12px;
-  color: var(--text-3);
-  text-align: center;
-  margin-bottom: 16px;
-  line-height: 1.5;
-}
-.manual { font-size: 13px; color: var(--text-3); }
-.manual a { color: var(--text-2); text-decoration: underline; }
-
-.error { font-size: 13px; color: var(--red); margin-bottom: 12px; }
-.error a { color: var(--red); text-decoration: underline; }
-
-/* 로딩 */
-.loadingBody {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  padding: 48px 24px 40px;
-}
-.eyebrow {
-  font-size: 11px;
-  font-weight: 700;
-  letter-spacing: 0.1em;
-  text-transform: uppercase;
-  color: var(--text-3);
-  margin-bottom: 10px;
-}
-.loadingTitle {
-  font-size: 30px;
-  font-weight: 800;
-  letter-spacing: -0.8px;
-  line-height: 1.2;
-  margin-bottom: 6px;
-}
-.loadingSub { font-size: 14px; color: var(--text-2); margin-bottom: 36px; }
-
-.stepList { display: flex; flex-direction: column; }
-.stepRow {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  padding: 13px 0;
-  border-bottom: 1px solid var(--border);
-  animation: stepIn 0.25s ease forwards;
-}
-@keyframes stepIn {
-  from { opacity: 0; transform: translateY(5px); }
-  to   { opacity: 1; transform: translateY(0); }
-}
-.stepVisible { opacity: 1; }
-
-.check {
-  width: 20px;
-  height: 20px;
-  border-radius: 50%;
-  border: 1.5px solid var(--border-2);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-shrink: 0;
-}
-.done { background: var(--green); border-color: var(--green); }
-.spinning { border-color: var(--amber); border-top-color: transparent; animation: spin 0.7s linear infinite; }
-@keyframes spin { to { transform: rotate(360deg); } }
-
-.stepName { font-size: 14px; font-weight: 500; }
-
-.progWrap { margin-top: auto; padding-top: 32px; }
-.progTrack { height: 2px; background: var(--border); border-radius: 2px; overflow: hidden; margin-bottom: 10px; }
-.progFill { height: 100%; background: var(--navy-mid); border-radius: 2px; transition: width 0.35s ease; }
-.progLabel { font-size: 12px; color: var(--text-3); }
