--- src/app/signals/page.module.css
@@ -1,211 +1,0 @@
-.wrap {
-  min-height: 100dvh;
-  background:
-    radial-gradient(circle at top left, rgba(255, 255, 255, 0.9), transparent 34%),
-    linear-gradient(180deg, #eef1f7 0%, #e8ecf6 100%);
-}
-
-.header {
-  padding: 18px 20px 10px;
-  color: #111827;
-}
-
-.eyebrow {
-  font-size: 13px;
-  font-weight: 700;
-  color: #9aa0aa;
-}
-
-.title {
-  margin-top: 8px;
-  font-size: 28px;
-  line-height: 1.05;
-  font-weight: 900;
-  letter-spacing: -0.05em;
-}
-
-.sub {
-  margin-top: 14px;
-  max-width: 560px;
-  font-size: 12px;
-  line-height: 1.6;
-  color: #a8adb7;
-}
-
-.metaRow {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px;
-  margin-top: 12px;
-}
-
-.metaBadge,
-.metaHint {
-  border-radius: 999px;
-  padding: 9px 12px;
-  font-size: 12px;
-  font-weight: 700;
-}
-
-.metaBadge {
-  background: rgba(255, 255, 255, 0.8);
-  color: var(--navy);
-  border: 1px solid rgba(26, 35, 126, 0.08);
-}
-
-.metaHint {
-  background: rgba(255, 255, 255, 0.55);
-  color: #8e95a3;
-}
-
-.body {
-  padding: 0 16px 112px;
-}
-
-.actions {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 10px;
-  margin-top: 18px;
-  margin-bottom: 8px;
-}
-
-.primaryButton,
-.secondaryButton,
-.inlineRetry {
-  border-radius: var(--radius-btn);
-  padding: 12px 16px;
-  font-size: 13px;
-  font-weight: 800;
-}
-
-.primaryButton {
-  background: var(--navy);
-  color: #fff;
-  border: none;
-}
-
-.secondaryButton {
-  background: rgba(255, 255, 255, 0.85);
-  color: var(--navy);
-  border: 1px solid rgba(26, 35, 126, 0.12);
-}
-
-.cardList {
-  display: grid;
-  gap: 14px;
-}
-
-.summaryCard {
-  margin-bottom: 16px;
-  padding: 18px 16px;
-  border: 1px solid rgba(15, 23, 42, 0.04);
-  border-radius: 18px;
-  background: rgba(255, 255, 255, 0.95);
-  box-shadow: 0 16px 42px rgba(15, 23, 42, 0.08);
-}
-
-.summaryHeader {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  gap: 12px;
-}
-
-.summaryTitle {
-  font-size: 16px;
-  line-height: 1.2;
-  letter-spacing: -0.04em;
-  color: #111827;
-}
-
-.summaryChips {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  margin-top: 14px;
-}
-
-.summaryTiles {
-  display: flex;
-  gap: 8px;
-  margin-top: 14px;
-}
-
-.summaryTile {
-  flex: 1;
-  padding: 18px 10px;
-  border-radius: 18px;
-  text-align: center;
-}
-
-.summaryTileValue {
-  font-size: 42px;
-  font-weight: 900;
-  letter-spacing: -0.04em;
-  line-height: 1;
-}
-
-.summaryTileLabel {
-  margin-top: 8px;
-  font-size: 12px;
-  font-weight: 700;
-  line-height: 1.35;
-  white-space: pre-line;
-}
-
-.summaryLegend {
-  margin-top: 18px;
-  padding: 14px 16px;
-  border-radius: 16px;
-  background: #f7f7fb;
-  font-size: 12px;
-  font-weight: 700;
-  line-height: 1.6;
-  color: #5f6571;
-}
-
-.stateCard {
-  background: rgba(255, 255, 255, 0.88);
-  border: 1px solid var(--border);
-  border-radius: 18px;
-  padding: 18px;
-  margin-bottom: 16px;
-  font-size: 14px;
-  line-height: 1.6;
-  color: var(--text-2);
-}
-
-.inlineRetry {
-  margin-top: 12px;
-  background: var(--surface);
-  color: var(--navy);
-  border: 1px solid rgba(26, 35, 126, 0.14);
-}
-
-@media (min-width: 768px) {
-  .body {
-    max-width: 920px;
-    margin: 0 auto;
-    padding: 0 24px 128px;
-  }
-
-  .cardList {
-    grid-template-columns: repeat(2, minmax(0, 1fr));
-    align-items: start;
-  }
-
-  .summaryCard {
-    padding: 22px;
-  }
-}
-
-@media (max-width: 479px) {
-  .summaryTiles {
-    flex-wrap: wrap;
-  }
-
-  .summaryTile {
-    min-width: calc(33.333% - 6px);
-  }
-}
