--- src/app/style/page.module.css
@@ -1,437 +1,0 @@
-/* src/app/style/page.module.css */
-.wrap { display: flex; flex-direction: column; min-height: 100dvh; }
-
-/* 네이비 헤더 */
-.header {
-  background: var(--navy);
-  padding: 20px 20px 28px;
-}
-
-.eyebrow {
-  font-size: 11px;
-  font-weight: 700;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-  color: rgba(255,255,255,.5);
-  margin-bottom: 12px;
-}
-
-.bigLabel {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  font-size: 36px;
-  font-weight: 800;
-  letter-spacing: -1.5px;
-  line-height: 1;
-  color: #fff;
-  margin-bottom: 10px;
-}
-
-.emoji { font-size: 32px; line-height: 1; }
-
-.desc {
-  font-size: 14px;
-  font-weight: 500;
-  color: rgba(255,255,255,.7);
-  line-height: 1.5;
-}
-
-/* 바디 */
-.body {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  padding: 20px 0 108px;
-}
-
-.sectionLabel {
-  font-size: 11px;
-  font-weight: 800;
-  letter-spacing: 0.1em;
-  text-transform: uppercase;
-  color: var(--text-3);
-  padding: 0 20px 12px;
-}
-
-.quizResult {
-  margin: 0 20px 12px;
-  padding: 14px 16px;
-  border-radius: var(--radius-card);
-  background: #eff6ff;
-  border: 1px solid #bfdbfe;
-  font-size: 14px;
-  font-weight: 700;
-  color: #1d4ed8;
-}
-
-.quizEntry {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 16px;
-  margin: 0 20px 16px;
-  padding: 18px;
-  border: 1px solid var(--border);
-  border-radius: var(--radius-card);
-  background: linear-gradient(135deg, #fff, #f8fafc);
-}
-
-.quizEntryTitle {
-  display: block;
-  font-size: 16px;
-  font-weight: 800;
-  color: var(--text-1);
-}
-
-.quizEntryDesc {
-  margin: 6px 0 0;
-  font-size: 13px;
-  line-height: 1.45;
-  color: var(--text-2);
-}
-
-.quizEntryButton {
-  flex-shrink: 0;
-  border: 0;
-  border-radius: 999px;
-  padding: 12px 16px;
-  background: var(--navy);
-  color: #fff;
-  font-size: 14px;
-  font-weight: 700;
-  cursor: pointer;
-}
-
-.quizWrap {
-  margin: 0 20px 16px;
-  padding: 20px;
-  border-radius: var(--radius-card);
-  background: var(--surface);
-  border: 1px solid var(--border);
-}
-
-.quizProgress {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 12px;
-  margin-bottom: 12px;
-}
-
-.quizProgressLabel {
-  font-size: 12px;
-  font-weight: 800;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-  color: var(--text-3);
-}
-
-.quizProgressStep {
-  font-size: 13px;
-  font-weight: 700;
-  color: var(--navy);
-}
-
-.quizQuestion {
-  margin: 0 0 16px;
-  font-size: 22px;
-  line-height: 1.4;
-  letter-spacing: -0.03em;
-  color: var(--text-1);
-}
-
-.quizChoices {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.quizChoice {
-  width: 100%;
-  padding: 16px;
-  border: 1px solid var(--border);
-  border-radius: 16px;
-  background: #fff;
-  text-align: left;
-  font-size: 15px;
-  line-height: 1.5;
-  color: var(--text-1);
-  cursor: pointer;
-  transition: border-color 0.15s, box-shadow 0.15s, transform 0.15s;
-}
-
-.quizChoiceSelected {
-  border-color: var(--navy);
-  box-shadow: 0 0 0 3px rgba(26,35,126,.08);
-  transform: translateY(-1px);
-}
-
-.quizNav {
-  display: flex;
-  justify-content: space-between;
-  gap: 12px;
-  margin-top: 18px;
-}
-
-.quizNavButton,
-.quizNavButtonPrimary {
-  min-width: 92px;
-  border-radius: 999px;
-  padding: 12px 16px;
-  font-size: 14px;
-  font-weight: 700;
-  cursor: pointer;
-}
-
-.quizNavButton {
-  border: 1px solid var(--border);
-  background: #fff;
-  color: var(--text-1);
-}
-
-.quizNavButtonPrimary {
-  margin-left: auto;
-  border: 0;
-  background: var(--navy);
-  color: #fff;
-}
-
-.quizNavButton:disabled,
-.quizNavButtonPrimary:disabled,
-.quizEntryButton:disabled {
-  opacity: 0.45;
-  cursor: not-allowed;
-}
-
-/* 동일성향 배너 */
-.sameBanner {
-  margin: 0 20px 12px;
-  background: var(--green-bg);
-  border: 1px solid #A7F3D0;
-  border-radius: var(--radius-card);
-  padding: 12px 16px;
-  font-size: 14px;
-  font-weight: 600;
-  color: var(--green);
-}
-
-.selectionRequired {
-  margin: 4px 20px 12px;
-  font-size: 13px;
-  font-weight: 600;
-  color: #b45309;
-}
-
-/* 카드 스크롤 래퍼 — 오른쪽 페이드 */
-.cardWrap {
-  position: relative;
-}
-.cardWrap::after {
-  content: '';
-  position: absolute;
-  top: 0;
-  right: 0;
-  bottom: 0;
-  width: 48px;
-  background: linear-gradient(to right, transparent, var(--bg));
-  pointer-events: none;
-}
-
-/* 카드 가로 스크롤 트랙 */
-.cardTrack {
-  display: flex;
-  flex-direction: row;
-  gap: 12px;
-  padding: 4px 0 8px;
-  overflow-x: auto;
-  scroll-padding-left: 20px;
-  scroll-snap-type: x mandatory;
-  -webkit-overflow-scrolling: touch;
-  scrollbar-width: none;
-}
-.cardTrack::-webkit-scrollbar { display: none; }
-.cardTrack::before {
-  content: '';
-  flex: 0 0 8px; /* 8px width + 12px gap = 20px left margin */
-}
-.cardTrack::after {
-  content: '';
-  flex: 0 0 36px; /* 36px width + 12px gap = 48px right margin */
-}
-
-/* 프리셋 카드 */
-.card {
-  flex: 0 0 min(220px, 72vw);
-  scroll-snap-align: start;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  gap: 6px;
-  padding: 20px 18px;
-  background: var(--surface);
-  border: 2px solid var(--border);
-  border-radius: var(--radius-card);
-  text-align: left;
-  cursor: pointer;
-  position: relative;
-  transition: border-color 0.15s, box-shadow 0.15s;
-}
-.card:active { opacity: 0.85; }
-
-.cardSelected {
-  border-color: var(--navy);
-  box-shadow: 0 0 0 3px rgba(26,35,126,.08);
-}
-
-/* 체크마크 */
-.check {
-  position: absolute;
-  top: 12px;
-  right: 14px;
-  font-size: 14px;
-  font-weight: 800;
-  color: var(--navy);
-  line-height: 1;
-}
-
-.cardEmoji { font-size: 28px; line-height: 1; margin-bottom: 4px; }
-
-.cardLabel {
-  font-size: 16px;
-  font-weight: 700;
-  letter-spacing: -0.3px;
-  color: var(--text-1);
-}
-
-.cardDesc {
-  font-size: 13px;
-  color: var(--text-2);
-  line-height: 1.45;
-}
-
-.cardAlloc {
-  font-size: 11px;
-  color: var(--text-3);
-  margin-top: 6px;
-  letter-spacing: 0.01em;
-}
-
-.partDivider {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  padding: 20px 16px 18px;
-}
-
-.partDividerLine {
-  flex: 1;
-  margin: 0;
-  border: 0;
-  border-top: 1px solid var(--border);
-}
-
-.partDividerLabel {
-  flex-shrink: 0;
-  font-size: 12px;
-  font-weight: 700;
-  letter-spacing: -0.01em;
-  color: var(--text-3);
-}
-
-.targetSection {
-  padding: 0 20px 108px;
-}
-
-.targetHeader {
-  margin-bottom: 12px;
-}
-
-.targetTitle {
-  margin: 0;
-  font-size: 18px;
-  font-weight: 800;
-  letter-spacing: -0.03em;
-  color: var(--text-1);
-}
-
-.targetHint {
-  margin: 6px 0 0;
-  font-size: 13px;
-  line-height: 1.5;
-  color: var(--text-2);
-}
-
-.sliderList {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.sliderRow {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  padding: 16px;
-  border: 1px solid var(--border);
-  border-radius: 16px;
-  background: #fff;
-}
-
-.sliderMeta {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 12px;
-}
-
-.sliderLabel {
-  font-size: 15px;
-  font-weight: 700;
-  color: var(--text-1);
-}
-
-.sliderValue {
-  font-size: 14px;
-  font-weight: 700;
-  color: var(--navy);
-}
-
-.slider {
-  width: 100%;
-  accent-color: var(--navy);
-}
-
-.targetSummary {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  margin-top: 12px;
-  padding: 14px 16px;
-  border-radius: 14px;
-  font-size: 14px;
-}
-
-.targetSummaryValid {
-  background: #ecfdf3;
-  color: #166534;
-}
-
-.targetSummaryInvalid {
-  background: #fef2f2;
-  color: #b91c1c;
-}
-
-.targetSummaryLabel {
-  font-weight: 600;
-}
-
-.targetSummaryValue {
-  font-size: 16px;
-}
-
-.targetSummaryMessage {
-  margin-left: auto;
-  text-align: right;
-}
-
