--- src/components/ActionItem.module.css
@@ -1,113 +1,0 @@
-/* src/components/ActionItem.module.css */
-.item { overflow: hidden; }
-.item + .item { border-top: 1px solid var(--border); }
-
-.top { padding: 14px 16px 0; }
-
-.stockRow {
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  margin-bottom: 12px;
-}
-
-.avatar {
-  width: 42px;
-  height: 42px;
-  border-radius: 12px;
-  flex-shrink: 0;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 15px;
-  font-weight: 800;
-}
-.avatarSell { background: var(--red-bg); color: var(--red); }
-.avatarBuy  { background: var(--green-bg); color: var(--green); }
-
-.stockInfo { flex: 1; min-width: 0; }
-.name { font-size: 15px; font-weight: 700; letter-spacing: -0.2px; }
-.ticker { font-size: 11px; color: var(--text-3); margin-top: 2px; }
-
-.actionBadge {
-  font-size: 12px;
-  font-weight: 700;
-  padding: 5px 10px;
-  border-radius: 99px;
-  flex-shrink: 0;
-  white-space: nowrap;
-}
-.sell { background: var(--red-bg); color: var(--red); }
-.buy  { background: var(--green-bg); color: var(--green); }
-
-.amountBox {
-  display: flex;
-  gap: 8px;
-  background: #f7f8fc;
-  border-radius: 12px;
-  padding: 11px 14px;
-}
-
-.amountItem {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  gap: 3px;
-}
-.amountItem + .amountItem {
-  border-left: 1px solid #e8e8ec;
-  padding-left: 8px;
-}
-
-.amountLabel { font-size: 10px; color: var(--text-3); font-weight: 500; }
-
-.amountValue {
-  font-size: 14px;
-  font-weight: 800;
-  font-variant-numeric: tabular-nums;
-  letter-spacing: -0.3px;
-}
-.amountSell { color: var(--red); }
-.amountBuy  { color: var(--green); }
-.amountValueNeutral {
-  font-size: 14px;
-  font-weight: 700;
-  color: var(--text-2);
-  font-variant-numeric: tabular-nums;
-}
-
-.whySection {
-  border-top: 1px solid #f2f2f5;
-  margin-top: 12px;
-}
-
-.whyToggle {
-  width: 100%;
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 12px 16px;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  font-size: 12px;
-  font-weight: 600;
-  color: var(--text-2);
-  font-family: inherit;
-}
-.whyToggle:active { background: var(--bg); }
-
-.whyArrow {
-  font-size: 12px;
-  color: var(--text-3);
-  transition: transform 0.2s;
-  display: inline-block;
-}
-.whyArrowOpen { transform: rotate(180deg); }
-
-.whyText {
-  padding: 0 16px 14px;
-  font-size: 12.5px;
-  color: var(--text-2);
-  line-height: 1.75;
-}
