--- src/components/AllocationBar.module.css
@@ -1,80 +1,0 @@
-/* src/components/AllocationBar.module.css */
-.wrap {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.row {
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-}
-
-.rowHeader {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 10px;
-}
-
-.label {
-  font-size: 12px;
-  font-weight: 500;
-  color: rgba(255, 255, 255, 0.7);
-}
-
-.meta {
-  display: flex;
-  align-items: center;
-  gap: 7px;
-}
-
-.pct {
-  font-size: 13px;
-  font-weight: 700;
-  font-variant-numeric: tabular-nums;
-  color: #fff;
-}
-
-.targetLabel {
-  border-radius: 99px;
-  background: transparent;
-  padding: 0;
-  font-size: 10px;
-  font-weight: 600;
-  white-space: nowrap;
-  color: rgba(255, 255, 255, 0.4);
-}
-
-.targetOver {
-  color: rgb(232, 168, 56);
-  background: rgba(232, 168, 56, 0.18);
-  padding: 2px 6px;
-}
-
-.track {
-  position: relative;
-  height: 5px;
-  border-radius: 99px;
-  background: rgba(255, 255, 255, 0.12);
-}
-
-.fill {
-  position: absolute;
-  left: 0;
-  top: 0;
-  height: 100%;
-  border-radius: 99px;
-  transition: width 1s cubic-bezier(0.4, 0, 0.2, 1);
-}
-
-.targetLine {
-  position: absolute;
-  top: -3px;
-  width: 1.5px;
-  height: 11px;
-  border-radius: 1px;
-  transform: translateX(-50%);
-  background: rgba(255, 255, 255, 0.4);
-}
