--- src/components/BottomNav.module.css
@@ -1,54 +1,0 @@
-.nav {
-  position: fixed;
-  left: 50%;
-  bottom: 0;
-  z-index: 100;
-  display: flex;
-  width: 100%;
-  max-width: 440px;
-  padding-bottom: env(safe-area-inset-bottom, 0);
-  border-top: 1px solid #ebebef;
-  background: rgba(255, 255, 255, 0.96);
-  backdrop-filter: blur(18px);
-  box-shadow: 0 -12px 32px rgba(15, 23, 42, 0.06);
-  transform: translateX(-50%);
-}
-
-.item {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  align-items: center;
-  gap: 3px;
-  padding: 10px 0 12px;
-  color: #b7bac2;
-}
-
-.itemActive {
-  color: var(--navy);
-}
-
-.icon {
-  width: 20px;
-  height: 20px;
-}
-
-.icon svg {
-  width: 20px;
-  height: 20px;
-  overflow: visible;
-}
-
-.label {
-  font-size: 10px;
-  font-weight: 700;
-  letter-spacing: 0.02em;
-}
-
-.dot {
-  width: 6px;
-  height: 6px;
-  margin-top: -1px;
-  border-radius: 50%;
-  background: var(--navy);
-}
