--- src/components/ConfirmCard.module.css
@@ -1,215 +1,0 @@
-/* src/components/ConfirmCard.module.css */
-.card {
-  background: var(--surface);
-  border: 1px solid var(--border);
-  border-radius: var(--radius-card);
-  overflow: hidden;
-  transition: opacity 0.2s, transform 0.2s;
-}
-
-.main { padding: 16px 16px 14px; }
-
-.row1 {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  margin-bottom: 12px;
-}
-
-.nameWrap { display: flex; align-items: center; gap: 7px; flex: 1; min-width: 0; }
-
-.name {
-  font-size: 16px;
-  font-weight: 700;
-  letter-spacing: -0.3px;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-}
-
-.dupBadge {
-  background: var(--amber-bg);
-  color: var(--amber);
-  border: 1px solid var(--amber-border);
-  font-size: 10px;
-  font-weight: 700;
-  padding: 2px 6px;
-  border-radius: var(--radius-badge);
-  white-space: nowrap;
-  flex-shrink: 0;
-}
-
-.deleteBtn {
-  background: none;
-  border: none;
-  color: var(--text-3);
-  font-size: 16px;
-  min-width: 44px;
-  min-height: 44px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 8px;
-  margin: -10px -10px -10px 0;
-  flex-shrink: 0;
-}
-.deleteBtn:active { background: var(--bg); }
-
-/* 2×2 그리드 */
-.grid {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: 1px;
-  background: var(--border);
-  border: 1px solid var(--border);
-  border-radius: 10px;
-  overflow: hidden;
-  margin-bottom: 12px;
-}
-.cell { background: var(--surface); padding: 10px 12px; }
-.cellLabel {
-  font-size: 10px;
-  font-weight: 700;
-  letter-spacing: 0.04em;
-  text-transform: uppercase;
-  color: var(--text-3);
-  margin-bottom: 3px;
-}
-.cellVal {
-  font-size: 14px;
-  font-weight: 700;
-  font-variant-numeric: tabular-nums;
-  letter-spacing: -0.3px;
-}
-.cellInput {
-  width: 100%;
-  font-size: 14px;
-  font-weight: 700;
-  font-variant-numeric: tabular-nums;
-  letter-spacing: -0.3px;
-  font-family: var(--font);
-  background: none;
-  border: none;
-  border-bottom: 1px solid transparent;
-  padding: 0;
-  color: var(--text-1);
-  outline: none;
-  min-width: 0;
-}
-.cellInput:focus {
-  border-bottom-color: var(--primary);
-}
-.secondary { font-size: 13px; font-weight: 600; color: var(--text-2); }
-
-/* 자산군 + 섹터 */
-.meta { display: flex; gap: 8px; align-items: center; }
-.select {
-  flex: 1;
-  font-family: var(--font);
-  font-size: 13px;
-  font-weight: 600;
-  padding: 9px 28px 9px 10px;
-  border: 1px solid var(--border-2);
-  border-radius: var(--radius-input);
-  background: var(--bg);
-  color: var(--text-1);
-  appearance: none;
-  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='11' height='11' viewBox='0 0 24 24' fill='none' stroke='%239CA3AF' stroke-width='2.5'%3E%3Cpath d='m6 9 6 6 6-6'/%3E%3C/svg%3E");
-  background-repeat: no-repeat;
-  background-position: right 10px center;
-}
-.sectorSelect {
-  flex: 1;
-  min-width: 0;
-  font-size: 12px;
-  font-weight: 600;
-  font-family: var(--font);
-  color: var(--text-2);
-}
-.sectorSelect:focus {
-  border-color: var(--primary);
-}
-
-/* 768px+ 테이블 로우 */
-.tr { border-bottom: 1px solid var(--border); }
-.tr:last-child { border-bottom: none; }
-.trDup { background: var(--amber-bg); }
-
-.tdName {
-  padding: 12px 16px 12px 0;
-  min-width: 120px;
-}
-.td { padding: 12px 12px 12px 0; vertical-align: middle; }
-.tdNum { padding: 12px 12px 12px 0; vertical-align: middle; font-variant-numeric: tabular-nums; }
-.tdPct { color: var(--text-2); font-size: 13px; white-space: nowrap; }
-.tdAction { padding: 12px 0; vertical-align: middle; }
-.tdNameInner { display: flex; align-items: center; gap: 6px; }
-
-.tdInput {
-  width: 100%;
-  min-width: 70px;
-  font-size: 13px;
-  font-weight: 600;
-  font-variant-numeric: tabular-nums;
-  font-family: var(--font);
-  background: none;
-  border: none;
-  border-bottom: 1px solid transparent;
-  padding: 0;
-  color: var(--text-1);
-  outline: none;
-}
-.tdInput:focus { border-bottom-color: var(--primary); }
-
-.selectSm {
-  font-family: var(--font);
-  font-size: 12px;
-  font-weight: 600;
-  padding: 5px 22px 5px 7px;
-  border: 1px solid var(--border-2);
-  border-radius: var(--radius-input);
-  background: var(--bg);
-  color: var(--text-1);
-  appearance: none;
-  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='10' height='10' viewBox='0 0 24 24' fill='none' stroke='%239CA3AF' stroke-width='2.5'%3E%3Cpath d='m6 9 6 6 6-6'/%3E%3C/svg%3E");
-  background-repeat: no-repeat;
-  background-position: right 7px center;
-}
-.sectorSelectSm {
-  min-width: 110px;
-  color: var(--text-2);
-}
-
-/* 상세보기 토글 */
-.toggle {
-  width: 100%;
-  background: none;
-  border: none;
-  border-top: 1px solid var(--border);
-  padding: 10px 16px;
-  font-size: 12px;
-  font-weight: 600;
-  color: var(--text-3);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 5px;
-}
-.arrowDown { display: inline-block; transition: transform 0.2s; }
-.arrowUp { display: inline-block; transform: rotate(180deg); transition: transform 0.2s; }
-
-/* 펼쳐진 상세 */
-.expand {
-  padding: 12px 16px 14px;
-  border-top: 1px solid var(--border);
-  background: var(--bg);
-}
-.expandRow {
-  display: flex;
-  justify-content: space-between;
-  font-size: 13px;
-  padding: 5px 0;
-}
-.expandRow + .expandRow { border-top: 1px solid var(--border); }
-.expandLabel { color: var(--text-2); }
-.expandVal { font-weight: 600; font-variant-numeric: tabular-nums; }
