--- src/components/ImprovementSheet.module.css
@@ -1,207 +1,0 @@
-.overlay {
-  position: fixed;
-  inset: 0;
-  background: rgba(15, 23, 42, 0.5);
-  z-index: 30;
-  display: flex;
-  align-items: flex-end;
-  justify-content: center;
-  padding-top: 24px;
-}
-
-.sheet {
-  width: 100%;
-  max-height: 80vh;
-  overflow-y: auto;
-  background: var(--surface);
-  border-radius: 24px 24px 0 0;
-  padding: 18px 18px calc(22px + env(safe-area-inset-bottom));
-  box-shadow: 0 -12px 32px rgba(15, 23, 42, 0.16);
-  animation: slideUp 0.22s ease-out;
-}
-
-.header {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  gap: 16px;
-  margin-bottom: 18px;
-}
-
-.eyebrow {
-  font-size: 11px;
-  font-weight: 800;
-  letter-spacing: 0.08em;
-  text-transform: uppercase;
-  color: var(--text-3);
-  margin-bottom: 4px;
-}
-
-.title {
-  font-size: 22px;
-  font-weight: 800;
-  letter-spacing: -0.6px;
-  color: var(--text-1);
-}
-
-.closeButton {
-  width: 36px;
-  height: 36px;
-  border: 1px solid var(--border);
-  border-radius: 50%;
-  background: var(--surface);
-  color: var(--text-2);
-  font-size: 18px;
-  line-height: 1;
-  flex-shrink: 0;
-}
-
-.scoreCard {
-  display: grid;
-  grid-template-columns: 1fr auto 1fr;
-  gap: 10px;
-  align-items: center;
-  padding: 18px;
-  border-radius: 18px;
-  background: linear-gradient(135deg, var(--navy-tint), #f8faff);
-  margin-bottom: 20px;
-}
-
-.scoreBlock {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-
-.scoreLabel {
-  font-size: 12px;
-  font-weight: 700;
-  color: var(--text-3);
-}
-
-.scoreValue {
-  font-size: 30px;
-  font-weight: 800;
-  letter-spacing: -1px;
-  color: var(--navy);
-}
-
-.scoreValueIdeal { color: var(--green); }
-
-.scoreArrow {
-  font-size: 24px;
-  font-weight: 700;
-  color: rgba(26, 35, 126, 0.5);
-}
-
-.section + .section { margin-top: 22px; }
-
-.sectionTitle {
-  font-size: 14px;
-  font-weight: 800;
-  color: var(--text-1);
-  margin-bottom: 12px;
-}
-
-.rowList {
-  display: flex;
-  flex-direction: column;
-  gap: 12px;
-}
-
-.row {
-  padding: 14px;
-  border: 1px solid var(--border);
-  border-radius: 16px;
-  background: #fbfcff;
-}
-
-.rowTop {
-  display: flex;
-  align-items: baseline;
-  justify-content: space-between;
-  gap: 10px;
-  margin-bottom: 10px;
-}
-
-.bucketLabel {
-  font-size: 14px;
-  font-weight: 700;
-  color: var(--text-1);
-}
-
-.rowMeta {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-  flex-wrap: wrap;
-  justify-content: flex-end;
-}
-
-.percent {
-  font-size: 14px;
-  font-weight: 800;
-  color: var(--text-1);
-  font-variant-numeric: tabular-nums;
-}
-
-.deltaNeutral {
-  font-size: 13px;
-  font-weight: 700;
-  color: var(--text-3);
-}
-
-.deltaAlert {
-  font-size: 13px;
-  font-weight: 700;
-  color: var(--red);
-}
-
-.targetCheck {
-  font-size: 14px;
-  font-weight: 800;
-  color: var(--green);
-}
-
-.track {
-  width: 100%;
-  height: 10px;
-  background: #e8edf8;
-  border-radius: 999px;
-  overflow: hidden;
-}
-
-.trackFill {
-  height: 100%;
-  border-radius: 999px;
-  background: linear-gradient(90deg, var(--navy), var(--navy-mid));
-}
-
-.trackFillTarget {
-  background: linear-gradient(90deg, #10b981, #34d399);
-}
-
-@keyframes slideUp {
-  from {
-    transform: translateY(24px);
-    opacity: 0;
-  }
-  to {
-    transform: translateY(0);
-    opacity: 1;
-  }
-}
-
-@media (min-width: 768px) {
-  .overlay {
-    align-items: center;
-    padding: 24px;
-  }
-
-  .sheet {
-    max-width: 480px;
-    max-height: min(80vh, 720px);
-    border-radius: 24px;
-    box-shadow: 0 20px 48px rgba(15, 23, 42, 0.2);
-  }
-}
