--- src/components/MarketBanner.module.css
@@ -1,96 +1,0 @@
-.banner {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  padding: 18px;
-  border-radius: 22px;
-  background:
-    radial-gradient(circle at top right, rgba(255, 255, 255, 0.16), transparent 42%),
-    linear-gradient(135deg, var(--navy-dark), var(--navy));
-  color: #fff;
-  box-shadow: 0 18px 34px rgba(18, 24, 88, 0.18);
-}
-
-.copy {
-  display: grid;
-  gap: 10px;
-}
-
-.eyebrow {
-  font-size: 11px;
-  font-weight: 800;
-  letter-spacing: 0.12em;
-  text-transform: uppercase;
-  color: rgba(255, 255, 255, 0.68);
-}
-
-.titleRow {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  gap: 12px;
-}
-
-.title {
-  font-size: 22px;
-  font-weight: 900;
-  letter-spacing: -0.04em;
-}
-
-.badge {
-  flex-shrink: 0;
-  border-radius: 999px;
-  padding: 8px 12px;
-  font-size: 12px;
-  font-weight: 800;
-}
-
-.badge_risk_on {
-  background: rgba(5, 150, 105, 0.16);
-  color: #d8fff1;
-}
-
-.badge_neutral {
-  background: rgba(217, 119, 6, 0.16);
-  color: #ffefcf;
-}
-
-.badge_risk_off {
-  background: rgba(220, 38, 38, 0.16);
-  color: #ffd6d6;
-}
-
-.badge_loading {
-  background: rgba(255, 255, 255, 0.12);
-  color: rgba(255, 255, 255, 0.82);
-}
-
-.summary {
-  font-size: 14px;
-  line-height: 1.65;
-  color: rgba(255, 255, 255, 0.82);
-}
-
-.link {
-  display: inline-flex;
-  align-items: center;
-  justify-content: center;
-  min-height: 44px;
-  border-radius: 14px;
-  background: rgba(255, 255, 255, 0.12);
-  border: 1px solid rgba(255, 255, 255, 0.14);
-  font-size: 13px;
-  font-weight: 800;
-}
-
-@media (min-width: 768px) {
-  .banner {
-    flex-direction: row;
-    align-items: flex-end;
-    justify-content: space-between;
-  }
-
-  .link {
-    min-width: 180px;
-  }
-}
