--- src/components/MarketCard.module.css
@@ -1,198 +1,0 @@
-.card {
-  margin: 0 16px 8px;
-  overflow: hidden;
-  border-radius: 18px;
-  background: #fff;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.06);
-}
-
-.toggle {
-  width: 100%;
-  display: flex;
-  align-items: center;
-  gap: 12px;
-  border: none;
-  background: none;
-  padding: 14px 16px;
-  text-align: left;
-  font-family: inherit;
-}
-
-.copy {
-  min-width: 0;
-  flex: 1;
-}
-
-.iconWrap {
-  width: 36px;
-  height: 36px;
-  flex-shrink: 0;
-  border-radius: 10px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background: transparent;
-}
-
-.iconWrap_positive {
-  background: rgba(35, 178, 109, 0.07);
-}
-
-.iconWrap_negative {
-  background: rgba(226, 77, 77, 0.07);
-}
-
-.iconWrap_neutral,
-.iconWrap_unavailable {
-  background: transparent;
-}
-
-.icon {
-  display: inline-flex;
-  width: 18px;
-  height: 18px;
-}
-
-.icon svg {
-  width: 18px;
-  height: 18px;
-}
-
-.icon_positive {
-  color: #23b26d;
-}
-
-.icon_negative {
-  color: #e24d4d;
-}
-
-.icon_neutral,
-.icon_unavailable {
-  color: #aaa;
-}
-
-.title {
-  margin-bottom: 2px;
-  font-size: 13px;
-  font-weight: 700;
-  color: #111;
-}
-
-.description {
-  display: -webkit-box;
-  overflow: hidden;
-  font-size: 11.5px;
-  line-height: 1.45;
-  color: #888;
-  -webkit-box-orient: vertical;
-  -webkit-line-clamp: 1;
-}
-
-.right {
-  flex-shrink: 0;
-  display: flex;
-  flex-direction: column;
-  align-items: flex-end;
-  gap: 5px;
-}
-
-.badge {
-  border-radius: 999px;
-  padding: 3px 9px;
-  font-size: 11px;
-  font-weight: 700;
-}
-
-.badge_positive {
-  background: rgba(35, 178, 109, 0.07);
-  color: #23b26d;
-}
-
-.badge_negative {
-  background: rgba(226, 77, 77, 0.07);
-  color: #e24d4d;
-}
-
-.badge_neutral {
-  background: transparent;
-  color: #aaa;
-}
-
-.badge_unavailable {
-  background: transparent;
-  color: #bbb;
-}
-
-.arrow {
-  display: inline-block;
-  font-size: 10px;
-  color: #ccc;
-  transition: transform 0.25s;
-}
-
-.arrowOpen {
-  transform: rotate(180deg);
-}
-
-.body {
-  display: grid;
-  gap: 12px;
-  padding: 0 16px 16px;
-  border-top: 1px solid #f3f3f5;
-}
-
-.note {
-  border-radius: 12px;
-  padding: 10px 12px;
-  background: #f6f7fa;
-}
-
-.noteText {
-  font-size: 12px;
-  line-height: 1.65;
-  color: #666;
-}
-
-.detailToggle {
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  border: none;
-  background: none;
-  padding: 0;
-  font-family: inherit;
-  font-size: 11px;
-  color: #bbb;
-}
-
-.detailArrow {
-  display: inline-block;
-  width: 10px;
-}
-
-.metricBox {
-  border-radius: 12px;
-  padding: 10px 12px;
-  background: #f8f8fc;
-}
-
-.metricTitle {
-  margin-bottom: 4px;
-  font-size: 11px;
-  font-weight: 700;
-  color: #888;
-}
-
-.metricSource {
-  font-family: ui-monospace, SFMono-Regular, Menlo, Monaco, Consolas, 'Liberation Mono', monospace;
-  font-size: 11px;
-  line-height: 1.5;
-  color: #aaa;
-}
-
-.metricValue {
-  margin-top: 6px;
-  font-size: 12px;
-  font-weight: 700;
-  color: #555;
-}
