--- src/components/ProblemCard.module.css
@@ -1,111 +1,0 @@
-/* src/components/ProblemCard.module.css */
-.card {
-  border: 1px solid var(--border);
-  border-radius: 18px;
-  background: var(--surface);
-  padding: 18px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.06);
-}
-
-.badge {
-  display: inline-flex;
-  align-items: center;
-  gap: 5px;
-  margin-bottom: 8px;
-  border-radius: 999px;
-  padding: 3px 10px;
-  font-size: 10px;
-  font-weight: 700;
-  letter-spacing: 0.03em;
-}
-
-.badgeHigh {
-  background: #fef2f2;
-  color: #e6514a;
-}
-
-.badgeMedium {
-  background: #fff8ea;
-  color: #d97706;
-}
-
-.dot {
-  width: 6px;
-  height: 6px;
-  border-radius: 50%;
-  flex-shrink: 0;
-}
-
-.dotHigh {
-  background: #e6514a;
-}
-
-.dotMedium {
-  background: #d97706;
-}
-
-.title {
-  margin-bottom: 14px;
-  font-size: 17px;
-  font-weight: 800;
-  letter-spacing: -0.03em;
-  line-height: 1.3;
-  color: var(--text-1);
-}
-
-.metrics {
-  display: flex;
-  align-items: stretch;
-  gap: 8px;
-  margin-bottom: 12px;
-  border-radius: 14px;
-  background: #f7f8fc;
-  padding: 14px;
-}
-
-.metric {
-  display: flex;
-  flex: 1;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  gap: 4px;
-  text-align: center;
-}
-
-.metricLabel {
-  font-size: 11px;
-  font-weight: 500;
-  color: #aaa;
-}
-
-.metricValue {
-  font-size: 32px;
-  font-weight: 900;
-  line-height: 1;
-  letter-spacing: -0.04em;
-  font-variant-numeric: tabular-nums;
-  color: #bbb;
-}
-
-.currentRed {
-  color: #e24d4d;
-}
-
-.currentAmber {
-  color: #d97706;
-}
-
-.metricDivider {
-  width: 1px;
-  margin: 0;
-  background: #e8e8ec;
-  border-radius: 1px;
-}
-
-.desc {
-  font-size: 13px;
-  font-weight: 500;
-  line-height: 1.7;
-  color: #666;
-}
