--- src/components/SectorPieChart.module.css
@@ -1,137 +1,0 @@
-.card {
-  overflow: hidden;
-  border: 1px solid var(--border);
-  border-radius: 18px;
-  background: var(--surface);
-  padding: 18px;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.06);
-}
-
-.header {
-  display: flex;
-  align-items: flex-start;
-  justify-content: space-between;
-  gap: 12px;
-  margin-bottom: 4px;
-}
-
-.eyebrow {
-  margin-bottom: 3px;
-  font-size: 10px;
-  font-weight: 600;
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  color: #bbb;
-}
-
-.title {
-  font-size: 17px;
-  font-weight: 800;
-  letter-spacing: -0.03em;
-  color: #111;
-}
-
-.caption {
-  margin-bottom: 14px;
-  font-size: 12px;
-  line-height: 1.55;
-  color: #999;
-}
-
-.topSector {
-  flex-shrink: 0;
-  border-radius: 99px;
-  background: rgba(226, 77, 77, 0.063);
-  padding: 4px 10px;
-  font-size: 10px;
-  font-weight: 700;
-  color: #e24d4d;
-  white-space: nowrap;
-}
-
-.content {
-  display: flex;
-  align-items: center;
-  gap: 14px;
-}
-
-.chartWrap {
-  flex-shrink: 0;
-}
-
-.chart {
-  flex-shrink: 0;
-}
-
-.chartValue {
-  font-size: 12px;
-  font-weight: 800;
-  fill: #1c2b5e;
-}
-
-.chartLabel {
-  font-size: 8.5px;
-  fill: #999;
-}
-
-.legend {
-  flex: 1;
-}
-
-.legendItem {
-  display: flex;
-  align-items: center;
-  gap: 7px;
-  margin-bottom: 9px;
-}
-
-.legendItem:last-child {
-  margin-bottom: 0;
-}
-
-.legendDot {
-  width: 8px;
-  height: 8px;
-  border-radius: 2px;
-  background: var(--slice-color);
-  flex-shrink: 0;
-}
-
-.legendLabel {
-  flex: 1;
-  font-size: 12px;
-  font-weight: 500;
-  color: #444;
-}
-
-.legendMeta {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  margin-left: auto;
-}
-
-.legendShare {
-  min-width: 36px;
-  font-size: 12px;
-  font-weight: 700;
-  text-align: right;
-  color: #111;
-}
-
-.legendValue {
-  min-width: 42px;
-  font-size: 11px;
-  text-align: right;
-  color: #ccc;
-}
-
-@media (max-width: 420px) {
-  .content {
-    gap: 12px;
-  }
-
-  .legendMeta {
-    gap: 6px;
-  }
-}
